+++ tests/test_lab_dev/test_circuit_components_utils.py
@@ -20,12 +20,9 @@ import pytest
 import numpy as np
 
 from mrmustard import math
-from mrmustard.lab_dev.circuit_components_utils import DsMap, CftMap
+from mrmustard.lab_dev.circuit_components_utils import _DsMap
 from mrmustard.lab_dev.states.base import DM
-from mrmustard.physics.triples import (
-    displacement_map_s_parametrized_Abc,
-    complex_fourier_transform_Abc,
-)
+from mrmustard.physics.triples import displacement_map_s_parametrized_Abc
 from mrmustard.physics.bargmann import wigner_to_bargmann_rho
 from mrmustard.physics.gaussian_integrals import contract_two_Abc
 
@@ -40,19 +37,19 @@ class TestDsMap:
 
     @pytest.mark.parametrize("modes,s", zip(modes, s))
     def test_init(self, modes, s):
-        dsmap = DsMap(modes, s)
+        dsmap = _DsMap(modes, s)
 
-        assert dsmap.name == "DsMap"
+        assert dsmap.name == "_DsMap"
         assert dsmap.modes == [modes] if not isinstance(modes, list) else sorted(modes)
 
     def test_representation(self):
-        rep1 = DsMap(modes=[0], s=0).representation
+        rep1 = _DsMap(modes=[0], s=0).representation
         A_correct, b_correct, c_correct = displacement_map_s_parametrized_Abc(s=0, n_modes=1)
         assert math.allclose(rep1.A[0], A_correct)
         assert math.allclose(rep1.b[0], b_correct)
         assert math.allclose(rep1.c[0], c_correct)
 
-        rep2 = DsMap(modes=[5, 10], s=1).representation
+        rep2 = _DsMap(modes=[5, 10], s=1).representation
         A_correct, b_correct, c_correct = displacement_map_s_parametrized_Abc(s=1, n_modes=2)
         assert math.allclose(rep2.A[0], A_correct)
         assert math.allclose(rep2.b[0], b_correct)
@@ -67,7 +64,7 @@ class TestDsMap:
         state_bargmann_triple = (A, b, c)
 
         # get new triple by right shift
-        state_after = state >> DsMap(modes=[0], s=0)
+        state_after = state >> _DsMap(modes=[0], s=0)
         A1, b1, c1 = state_after.bargmann_triple
 
         # get new triple by contraction
@@ -95,35 +92,13 @@ class TestDsMap:
         state_bargmann_triple = (A, b, c)
 
         # get new triple by right shift
-        state_after = state >> DsMap(modes=[0, 1], s=0)
+        state_after = state >> _DsMap(modes=[0, 1], s=0)
         A1, b1, c1 = state_after.bargmann_triple
 
         # get new triple by contraction
         Ds_bargmann_triple = displacement_map_s_parametrized_Abc(s=0, n_modes=2)
         A2, b2, c2 = contract_two_Abc(
             state_bargmann_triple, Ds_bargmann_triple, idx1=[0, 1, 2, 3], idx2=[2, 3, 6, 7]
-        )
-
-        assert math.allclose(A1[0], A2)
-        assert math.allclose(b1[0], b2)
-        assert math.allclose(c1[0], c2)
-
-    def test_cftmap_contraction_with_state(self):
-        # The init state cov and means comes from the random state 'state = Gaussian(1) >> Dgate([0.2], [0.3])'
-        state_cov = np.array([[0.32210229, -0.99732956], [-0.99732956, 6.1926484]])
-        state_means = np.array([0.4, 0.6])
-        A, b, c = wigner_to_bargmann_rho(state_cov, state_means)
-        state = DM.from_bargmann(modes=[0], triple=(A, b, c))
-        state_bargmann_triple = (A, b, c)
-
-        # get new triple by right shift
-        state_after = state >> CftMap(modes=[0])
-        A1, b1, c1 = state_after.bargmann_triple
-
-        # get new triple by contraction
-        Cft_bargmann_triple = complex_fourier_transform_Abc(n_modes=1)
-        A2, b2, c2 = contract_two_Abc(
-            state_bargmann_triple, Cft_bargmann_triple, idx1=[0, 1], idx2=[2, 3]
         )
 
         assert math.allclose(A1[0], A2)